# cmd/am2server/openapi.yml
openapi: 3.0.3
info:
  contact:
      name: sqlc-http
      url: https://github.com/walterwanderley/sqlc-http
  description: Manage and share am2/am2data (SincoANN profile) files
  title: am2 Server
  version: 0.0.1
...
paths:
...
  /captures/{id}:
      patch:
          parameters:
              - in: path
                name: id
                schema:
                  format: int64
                  type: integer
          requestBody:
              content:
                  application/json:
                      schema:
                          properties:
                              demo_link:
                                  type: string
                              description:
                                  type: string
                              has_cab:
                                  type: boolean
                              name:
                                  type: string
                              type:
                                  type: string
                          type: object
                  application/x-www-form-urlencoded:
                      schema:
                          properties:
                              demo_link:
                                  type: string
                              description:
                                  type: string
                              has_cab:
                                  type: boolean
                              name:
                                  type: string
                              type:
                                  type: string
                          type: object
          responses:
              "200":
                  content:
                      application/json:
                          schema:
                              properties:
                                  last_insert_id:
                                      format: int64
                                      type: integer
                                  rows_affected:
                                      format: int64
                                      type: integer
                              type: object
                  description: OK
              default:
                  content:
                      text/plain:
                          schema:
                              type: string
                  description: Error message
          summary: UpdateCapture
          tags:
              - capture
  /captures/{id}/file:
      get:
          parameters:
              - in: path
                name: id
                schema:
                  format: int64
                  type: integer
          responses:
              "200":
                  content:
                      application/octet-stream:
                          schema:
                              format: binary
                              type: string
                  description: OK
              default:
                  content:
                      text/plain:
                          schema:
                              type: string
                  description: Error message
          summary: GetCaptureFile
          tags:
              - capture
...
  /captures:
    get:
      post:
          requestBody:
              content:
                  multipart/form-data:
                      schema:
                          properties:
                              data:
                                  format: binary
                                  type: string
                              demo_link:
                                  type: string
                              description:
                                  type: string
                              has_cab:
                                  type: boolean
                              name:
                                  type: string
                              type:
                                  type: string
                              user_id:
                                  format: int64
                                  type: integer
                          type: object
          responses:
              "200":
                  content:
                      application/json:
                          schema:
                              properties:
                                  last_insert_id:
                                      format: int64
                                      type: integer
                                  rows_affected:
                                      format: int64
                                      type: integer
                              type: object
                  description: OK
              default:
                  content:
                      text/plain:
                          schema:
                              type: string
                  description: Error message
          summary: AddCapture
          tags:
              - capture
      summary: SearchCaptures
      tags:
          - capture
      
      
      parameters:
        - name: arg
          in: query
          schema:
            type: string
        - name: offset
          in: query
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          schema:
            type: integer
            format: int64
      
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/captureSearchCapturesRow"
          
        "default":    
          description: Error message
          content:
            text/plain:
              schema:
                type: string  
...
components:
  schemas:
...
    captureSearchCapturesRow:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        downloads:
          type: integer
          format: int64
        has_cab:
          type: boolean
        type:
          type: string
        created_at:
          type: string
          format: date-time
        demo_link:
          type: string
...